[Docs] components-button.md code for OnButtonClick example (Lua and Cpp) (#2703)
* Update components-button.md

Syntax code in how to use buttons and button onclick event in cpp and Lua

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update components-button.md

Added some more context to compare the differences between ::Handler and ::MultiHandler

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update components-button.md

Added UiButtonNotificationBus::GetCurrentBusId() so it can track which button was clicked, when multihandler is chosen.

Left them commented out so the user can choose

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update components-button.md

Matching the descriptions, code and language with the other pages

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update components-button.md

markdown on the code: lua and cpp! so it is not white on black only

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update components-button.md

Adding numbers to the text and code (as comments) to explain the process

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update components-button.md

EntityId as parameter of onbuttonclick does not exist. Removed

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

---------

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>
Co-authored-by: Shauna Gordon <shauna@shaunagordon.com>