[Docs] Dynamic Layout content populated with cpp (#2707)
* - Dynamic Layout content populated with cpp
- the example expect another ui component like the dropdown to already be present.

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update dynamic-layout.md

markdown with cpp so code is not only white on black

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update dynamic-layout.md

link for lua example. This is part of the LyShineExamples

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update dynamic-layout.md

Code sample includes how to clone and then modify the text-option using a vector of strings.
Extracted the code from here, and adapted it to the needs of the example:
https://github.com/o3de/o3de/blob/f17b405699c55400a985b8741af2649c39f11315/Gems/LyShineExamples/Code/Source/UiTestScrollBoxDataProviderComponent.cpp#L61

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update dynamic-layout.md

Included a link to another page, UIInitializationBus.
The dynamic layout should always be populated after ALL necessary UI elements are loaded, otherwise it can lead to unexpected results. The link at the bottom of the page, will suggest/guide the user to continue the tutorial with the use of UiButtonNotificationBus

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update dynamic-layout.md

fix for the internal urls, with relative path.

Kept the absolute github link thought. Nothing to do here.
https://github.com/o3de/o3de/blob/f17b405699c55400a985b8741af2649c39f11315/Gems/LyShineExamples/Assets/UI/Scripts/LyShineExamples/Dropdown/MultiSelectionDropdown.lua#L29

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

* Update dynamic-layout.md

Aprove 2731 one first: https://github.com/o3de/o3de.org/pull/2731
It will include the missing image
Then this one, that references it from the source.

Or manually edit this PR...

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>

---------

Signed-off-by: mythrz <44119535+mythrz@users.noreply.github.com>